pyproject.toml
[project]
name = "py-med-management-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "flask>=3.1.3",
    "flask-wtf>=1.2.0",
    "wtforms[email]>=3.1.0",
    "httpx>=0.27.0",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "uvicorn>=0.41.0",
    "cryptography>=46.0.5",
    "python-jose[cryptography]>=3.3.0",
    "python-dotenv>=1.0.0",
]
